+++ test/commands/head/head_time_stamp_test.dart
@@ -49,7 +49,7 @@ void main() {
           await initGit(d);
 
           // Add an uncommitted file
-          await initUncommittedFile(d);
+          await addFileWithoutCommitting(d);
 
           // Getting the head commit message should throw
           await expectLater(